PrescriptionService/requirement.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
mysqlclient==2.2.0
pymysql==1.1.0
python-dotenv==1.0.1
requests==2.31.0
grpcio==1.73.0
grpcio-tools==1.73.0
flasgger==0.9.7.1
pika